+++ rclone_manager/__init__.py
@@ -1,3 +1,4 @@
 from rclone_manager.rclone import RClone
 from rclone_manager.task import RCloneTask
 from rclone_manager.job import Job
+from rclone_manager.schedule import RCloneSchedule