+++ rclone_manager/task.py
@@ -1,10 +1,12 @@
 import time
 from rclone_manager.job import Job
+from rclone_manager import schedule
 
 
 class RCloneTask:
-    def __init__(self, job: Job = None, is_enabled=True, *args, **kwargs):
+    def __init__(self, job: Job = None, schedule: 'schedule.RCloneSchedule' = None, is_enabled=True, *args, **kwargs):
         self.job = job or Job(*args, **kwargs)
+        self.schedule = schedule
         self.is_enabled = is_enabled
 
         self.__start_time = None
@@ -14,8 +16,11 @@ class RCloneTask:
         if not self.is_enabled:
             return
         self.__start_time = time.time()
-        job = self.job.run(task=self, *args, **kwargs)
-        return job
+        if self.schedule:
+            self.schedule.schedule_start(task=self, *args, **kwargs)
+        else:
+            self.job.run(task=self, *args, **kwargs)
+        return self.job
 
     def result_handler(self, job: Job):
         self.__end_time = time.time()